Cleanup redundant type casts

+++ web/src/main/java/org/springframework/security/web/savedrequest/FastHttpDateFormat.java
@@ -177,7 +177,7 @@ public class FastHttpDateFormat {
 		Long cachedDate = null;
 
 		try {
-			cachedDate = (Long) parseCache.get(value);
+			cachedDate = parseCache.get(value);
 		}
 		catch (Exception ignored) {
 		}
